+++ core/opcodeCompiler/compiler/opCodeProcessor.go
@@ -153,14 +153,6 @@ func processByteCodes(code []byte) ([]byte, error) {
 	return DoCFGBasedOpcodeFusion(code)
 }
 
-func doOpcodesProcess(code []byte) ([]byte, error) {
-	code, err := doCodeFusion(code)
-	if err != nil {
-		return nil, ErrFailPreprocessing
-	}
-	return code, nil
-}
-
 // Exported version of doCodeFusion for use in benchmarks and external tests
 func DoCodeFusion(code []byte) ([]byte, error) {
 	// return doCodeFusion(code)
@@ -444,11 +436,18 @@ func applyFusionPatterns(code []byte, cur int, endPC int) int {
 
 		// Test zero and Jump. target offset at code[2-3]
 		if code0 == ISZERO && code1 == PUSH2 && code4 == JUMPI {
-			op := JumpIfZero
-			code[cur] = byte(op)
-			code[cur+1] = byte(Nop)
-			code[cur+4] = byte(Nop)
-			return 4
+			// Extract jump target from PUSH2 data (bytes 2-3)
+			if cur+3 < len(code) {
+				target := extractPush2Value(code[cur+2 : cur+4])
+				// Only do code fusion if the jump target is valid
+				if isValidJumpTarget(code, target) {
+					op := JumpIfZero
+					code[cur] = byte(op)
+					code[cur+1] = byte(Nop)
+					code[cur+4] = byte(Nop)
+					return 4
+				}
+			}
 		}
 
 		if code0 == DUP2 && code1 == MSTORE && code2 == PUSH1 && code4 == ADD {
@@ -507,17 +506,31 @@ func applyFusionPatterns(code []byte, cur int, endPC int) int {
 
 		// push and jump
 		if code0 == PUSH2 && code3 == JUMP {
-			op := Push2Jump
-			code[cur] = byte(op)
-			code[cur+3] = byte(Nop)
-			return 3
+			// Extract jump target from PUSH2 data (bytes 1-2)
+			if cur+2 < len(code) {
+				target := extractPush2Value(code[cur+1 : cur+3])
+				// Only do code fusion if the jump target is valid
+				if isValidJumpTarget(code, target) {
+					op := Push2Jump
+					code[cur] = byte(op)
+					code[cur+3] = byte(Nop)
+					return 3
+				}
+			}
 		}
 
 		if code0 == PUSH2 && code3 == JUMPI {
-			op := Push2JumpI
-			code[cur] = byte(op)
-			code[cur+3] = byte(Nop)
-			return 3
+			// Extract jump target from PUSH2 data (bytes 1-2)
+			if cur+2 < len(code) {
+				target := extractPush2Value(code[cur+1 : cur+3])
+				// Only do code fusion if the jump target is valid
+				if isValidJumpTarget(code, target) {
+					op := Push2JumpI
+					code[cur] = byte(op)
+					code[cur+3] = byte(Nop)
+					return 3
+				}
+			}
 		}
 
 		if code0 == PUSH1 && code2 == PUSH1 {
@@ -660,426 +673,6 @@ func getBlockType(block BasicBlock, blocks []BasicBlock, blockIndex int) string 
 	return "others"
 }
 
-func doCodeFusion(code []byte) ([]byte, error) {
-	fusedCode := make([]byte, len(code))
-	length := copy(fusedCode, code)
-	skipToNext := false
-	for i := 0; i < length; i++ {
-		cur := i
-		skipToNext = false
-
-		// todo: perf issue found with these logic, comment for now
-		//if fusedCode[cur] == byte(INVALID) {
-		//	return fusedCode, nil
-		//}
-		//if fusedCode[cur] >= minOptimizedOpcode && fusedCode[cur] <= maxOptimizedOpcode {
-		//	return code, ErrFailPreprocessing
-		//}
-
-		if length > cur+15 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code2 := ByteCode(fusedCode[cur+2])
-			code3 := ByteCode(fusedCode[cur+3])
-			code5 := ByteCode(fusedCode[cur+5])
-			code6 := ByteCode(fusedCode[cur+6])
-			code7 := ByteCode(fusedCode[cur+7])
-			code12 := ByteCode(fusedCode[cur+12])
-			code13 := ByteCode(fusedCode[cur+13])
-
-			if code0 == PUSH1 && code2 == CALLDATALOAD && code3 == PUSH1 && code5 == SHR &&
-				code6 == DUP1 && code7 == PUSH4 && code12 == GT && code13 == PUSH2 {
-				op := Push1CalldataloadPush1ShrDup1Push4GtPush2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+5] = byte(Nop)
-				fusedCode[cur+6] = byte(Nop)
-				fusedCode[cur+7] = byte(Nop)
-				fusedCode[cur+12] = byte(Nop)
-				fusedCode[cur+13] = byte(Nop)
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i += 15
-				continue
-			}
-		}
-
-		if length > cur+12 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-			code3 := ByteCode(fusedCode[cur+3])
-			code4 := ByteCode(fusedCode[cur+4])
-			code5 := ByteCode(fusedCode[cur+5])
-			code6 := ByteCode(fusedCode[cur+6])
-			code7 := ByteCode(fusedCode[cur+7])
-			code8 := ByteCode(fusedCode[cur+8])
-			code9 := ByteCode(fusedCode[cur+9])
-			code10 := ByteCode(fusedCode[cur+10])
-			code11 := ByteCode(fusedCode[cur+11])
-			code12 := ByteCode(fusedCode[cur+12])
-
-			if code0 == SWAP1 && code1 == PUSH1 && code3 == DUP1 && code4 == NOT &&
-				code5 == SWAP2 && code6 == ADD && code7 == AND && code8 == DUP2 &&
-				code9 == ADD && code10 == SWAP1 && code11 == DUP2 && code12 == LT {
-				op := Swap1Push1Dup1NotSwap2AddAndDup2AddSwap1Dup2LT
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				fusedCode[cur+5] = byte(Nop)
-				fusedCode[cur+6] = byte(Nop)
-				fusedCode[cur+7] = byte(Nop)
-				fusedCode[cur+8] = byte(Nop)
-				fusedCode[cur+9] = byte(Nop)
-				fusedCode[cur+10] = byte(Nop)
-				fusedCode[cur+11] = byte(Nop)
-				fusedCode[cur+12] = byte(Nop)
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i += 12
-				continue
-			}
-		}
-
-		if length > cur+9 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-			code2 := ByteCode(fusedCode[cur+2])
-			code3 := ByteCode(fusedCode[cur+2])
-			code4 := ByteCode(fusedCode[cur+2])
-			code5 := ByteCode(fusedCode[cur+2])
-			code6 := ByteCode(fusedCode[cur+6])
-			code7 := ByteCode(fusedCode[cur+7])
-
-			if code0 == DUP1 && code1 == PUSH4 && code6 == EQ && code7 == PUSH2 {
-				op := Dup1Push4EqPush2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+6] = byte(Nop)
-				fusedCode[cur+7] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP2 && code1 == SWAP1 && code2 == DUP3 && code3 == SUB && code4 == SWAP2 && code5 == DUP3 && code6 == GT && code7 == PUSH2 {
-				op := Swap2Swap1Dup3SubSwap2Dup3GtPush2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				fusedCode[cur+5] = byte(Nop)
-				fusedCode[cur+6] = byte(Nop)
-				fusedCode[cur+7] = byte(Nop)
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i += 9
-				continue
-			}
-		}
-
-		if length > cur+7 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code2 := ByteCode(fusedCode[cur+2])
-			code4 := ByteCode(fusedCode[cur+4])
-			code6 := ByteCode(fusedCode[cur+6])
-			code7 := ByteCode(fusedCode[cur+7])
-
-			if code0 == PUSH1 && code2 == PUSH1 && code4 == PUSH1 && code6 == SHL && code7 == SUB {
-				op := Push1Push1Push1SHLSub
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				fusedCode[cur+6] = byte(Nop)
-				fusedCode[cur+7] = byte(Nop)
-				skipToNext = true
-			}
-			if skipToNext {
-				i += 7
-				continue
-			}
-		}
-
-		if length > cur+5 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-			code2 := ByteCode(fusedCode[cur+2])
-			code3 := ByteCode(fusedCode[cur+3])
-			code4 := ByteCode(fusedCode[cur+4])
-			code5 := ByteCode(fusedCode[cur+5])
-
-			if code0 == AND && code1 == DUP2 && code2 == ADD && code3 == SWAP1 && code4 == DUP2 && code5 == LT {
-				op := AndDup2AddSwap1Dup2LT
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				fusedCode[cur+5] = byte(Nop)
-				skipToNext = true
-			}
-			if code0 == SUB && code1 == SLT && code2 == ISZERO && code3 == PUSH2 {
-				op := SubSLTIsZeroPush2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-			if code0 == DUP11 && code1 == MUL && code2 == DUP3 && code3 == SUB && code4 == MUL && code5 == DUP1 {
-				op := Dup11MulDup3SubMulDup1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				fusedCode[cur+5] = byte(Nop)
-				skipToNext = true
-			}
-			if skipToNext {
-				i += 5
-				continue
-			}
-		}
-
-		if length > cur+4 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-			code2 := ByteCode(fusedCode[cur+2])
-			code3 := ByteCode(fusedCode[cur+3])
-			code4 := ByteCode(fusedCode[cur+4])
-			if code0 == AND && code1 == SWAP1 && code2 == POP && code3 == SWAP2 && code4 == SWAP1 {
-				op := AndSwap1PopSwap2Swap1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-				skipToNext = true
-			}
-
-			// Test zero and Jump. target offset at code[2-3]
-			if code0 == ISZERO && code1 == PUSH2 && code4 == JUMPI {
-				op := JumpIfZero
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-
-				skipToNext = true
-			}
-
-			if code0 == DUP2 && code1 == MSTORE && code2 == PUSH1 && code4 == ADD {
-				op := Dup2MStorePush1Add
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-
-				skipToNext = true
-			}
-
-			if code0 == SHR && code1 == SHR && code2 == DUP1 && code3 == MUL && code4 == DUP1 {
-				op := SHRSHRDup1MulDup1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				fusedCode[cur+4] = byte(Nop)
-
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i += 4
-				continue
-			}
-		}
-
-		if length > cur+3 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-			code2 := ByteCode(fusedCode[cur+2])
-			code3 := ByteCode(fusedCode[cur+3])
-			if code0 == SWAP2 && code1 == SWAP1 && code2 == POP && code3 == JUMP {
-				op := Swap2Swap1PopJump
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP1 && code1 == POP && code2 == SWAP2 && code3 == SWAP1 {
-				op := Swap1PopSwap2Swap1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == POP && code1 == SWAP2 && code2 == SWAP1 && code3 == POP {
-				op := PopSwap2Swap1Pop
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-			// push and jump
-			if code0 == PUSH2 && code3 == JUMP {
-				op := Push2Jump
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == PUSH2 && code3 == JUMPI {
-				op := Push2JumpI
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == PUSH1 && code2 == PUSH1 {
-				op := Push1Push1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+2] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == ISZERO && code1 == PUSH2 {
-				op := IsZeroPush2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP3 && code1 == POP && code2 == POP && code3 == POP {
-				op := Swap3PopPopPop
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				fusedCode[cur+2] = byte(Nop)
-				fusedCode[cur+3] = byte(Nop)
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i += 3
-				continue
-			}
-		}
-
-		if length > cur+2 {
-			code0 := ByteCode(fusedCode[cur+0])
-			_ = ByteCode(fusedCode[cur+1])
-			code2 := ByteCode(fusedCode[cur+2])
-			if code0 == PUSH1 {
-				if code2 == ADD {
-					op := Push1Add
-					fusedCode[cur] = byte(op)
-					fusedCode[cur+2] = byte(Nop)
-					skipToNext = true
-				}
-				if code2 == SHL {
-					op := Push1Shl
-					fusedCode[cur] = byte(op)
-					fusedCode[cur+2] = byte(Nop)
-					skipToNext = true
-				}
-
-				if code2 == DUP1 {
-					op := Push1Dup1
-					fusedCode[cur] = byte(op)
-					fusedCode[cur+2] = byte(Nop)
-					skipToNext = true
-				}
-			}
-			if skipToNext {
-				i += 2
-				continue
-			}
-		}
-
-		if length > cur+1 {
-			code0 := ByteCode(fusedCode[cur+0])
-			code1 := ByteCode(fusedCode[cur+1])
-
-			if code0 == SWAP1 && code1 == POP {
-				op := Swap1Pop
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-			if code0 == POP && code1 == JUMP {
-				op := PopJump
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == POP && code1 == POP {
-				op := Pop2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP2 && code1 == SWAP1 {
-				op := Swap2Swap1
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP2 && code1 == POP {
-				op := Swap2Pop
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == DUP2 && code1 == LT {
-				op := Dup2LT
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == DUP3 && code1 == AND {
-				op := Dup3And
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if code0 == SWAP1 && code1 == DUP2 {
-				op := Swap1Dup2
-				fusedCode[cur] = byte(op)
-				fusedCode[cur+1] = byte(Nop)
-				skipToNext = true
-			}
-
-			if skipToNext {
-				i++
-				continue
-			}
-		}
-
-		skip, steps := calculateSkipSteps(fusedCode, cur)
-		if skip {
-			i += steps
-			continue
-		}
-	}
-	return fusedCode, nil
-}
-
 func calculateSkipSteps(code []byte, cur int) (skip bool, steps int) {
 	inst := ByteCode(code[cur])
 	if inst >= PUSH1 && inst <= PUSH32 {
@@ -1211,4 +804,99 @@ func isBlockTerminator(op ByteCode) bool {
 	default:
 		return false
 	}
+}
+
+// extractPush2Value extracts a 16-bit value from PUSH2 data bytes, consistent with EVM behavior
+func extractPush2Value(data []byte) uint64 {
+	if len(data) < 2 {
+		return 0
+	}
+	// EVM uses big-endian encoding for PUSH2 data
+	return uint64(data[0])<<8 | uint64(data[1])
+}
+
+// isValidJumpTarget checks if the given position is a valid jump destination.
+// A valid jump destination must be a JUMPDEST opcode and must be in a code segment (not data).
+func isValidJumpTarget(code []byte, target uint64) bool {
+	// Check bounds
+	if target >= uint64(len(code)) {
+		return false
+	}
+
+	// Check if the target is a JUMPDEST opcode
+	if ByteCode(code[target]) != JUMPDEST {
+		return false
+	}
+
+	// Check if the target is in a code segment (not data)
+	return isCodeSegment(code, target)
+}
+
+// isCodeSegment checks if the given position is in a code segment (not data).
+// This is a simplified version of the VM's code bitmap analysis.
+func isCodeSegment(code []byte, pos uint64) bool {
+	var pc uint64
+	for pc < uint64(len(code)) && pc <= pos {
+		op := ByteCode(code[pc])
+		pc++
+
+		// Handle super instructions that contain data
+		step, processed := codeBitmapForSI(code, pc, op)
+		if processed {
+			pc += step
+			continue
+		}
+
+		// Handle PUSH instructions (mark data bytes)
+		if op >= PUSH1 && op <= PUSH32 {
+			numbits := uint64(op - PUSH1 + 1)
+			// If the target position is within the data bytes of this PUSH, it's not code
+			if pos >= pc && pos < pc+numbits {
+				return false
+			}
+			pc += numbits
+		}
+	}
+
+	// If we reach here, the position is in a code segment
+	return true
+}
+
+// codeBitmapForSI is a simplified version for jump target validation
+func codeBitmapForSI(code []byte, pc uint64, op ByteCode) (step uint64, processed bool) {
+	switch op {
+	case Push2Jump, Push2JumpI:
+		step = 3
+		processed = true
+	case Push1Push1:
+		step = 3
+		processed = true
+	case Push1Add, Push1Shl, Push1Dup1:
+		step = 2
+		processed = true
+	case JumpIfZero:
+		step = 4
+		processed = true
+	case IsZeroPush2:
+		step = 3
+		processed = true
+	case Dup2MStorePush1Add:
+		step = 4
+		processed = true
+	case Dup1Push4EqPush2:
+		step = 9
+		processed = true
+	case Push1CalldataloadPush1ShrDup1Push4GtPush2:
+		step = 15
+		processed = true
+	case Push1Push1Push1SHLSub:
+		step = 7
+		processed = true
+	case Swap1Push1Dup1NotSwap2AddAndDup2AddSwap1Dup2LT:
+		step = 12
+		processed = true
+	default:
+		return 0, false
+	}
+	return step, processed
 }
